main_server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
structlog==23.2.0
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.8